Fix import + update test
